Add integration test for exception-lower post-lowering shape
Restructures testing to match the repo's intTests convention:

* New intTests/test_exception_lower/ verifies that SAW can load and
  verify the post-lowering pattern (thread-local error flag + sentinel
  return) on a hand-written 'golden' C++ source.  The test depends only
  on $SAW and a committed .bc artefact, so it runs in CI without the
  exception-lower binary or LLVM dev headers.

* saw-tools/exception-lower/test/ is retitled as a developer-only
  harness for working on the pass itself (compiles fixtures and runs
  the pass on them).  SAW verification is no longer duplicated here.

* error-return-value.cpp and verify-lowered.saw moved into intTests
  (entry points are now extern "C" for ABI-independent symbol names).
