+++ intTests/test_exception_lower/error-return-value.bc.FROM
@@ -1,0 +1,1 @@
+clang version 22.1.3 (https://github.com/llvm/llvm-project e9846648fd6183ee6d8cbdb4502213fcf902a211)
