+++ intTests/test_exception_lower/test.sh
@@ -1,0 +1,1 @@
+$SAW test.saw
