+++ saw-tools/exception-lower/test/README.md
@@ -1,8 +1,15 @@
-# exception-lower test suite
+# exception-lower developer harness
 
-Integration tests for the `exception-lower` LLVM pass.  Each test case
-exercises a different C++ exception-handling pattern and verifies that the
-lowering transformation produces correct error-return control flow.
+Developer-facing fixtures for the `exception-lower` LLVM pass.  Each
+C++ source here exercises a different EH pattern.  The shell harness
+builds them, runs the pass, and (optionally) disassembles the output
+for inspection.
+
+This directory is **not** part of the saw-script integration-test suite;
+it is intended for developers actively working on the pass.  The
+post-lowering shape is verified end-to-end by the integration test
+[`intTests/test_exception_lower/`](../../../intTests/test_exception_lower/),
+which does not require the pass binary to be built.
 
 ## Test cases
 
@@ -12,25 +19,12 @@ lowering transformation produces correct error-return control flow.
 | `nested-try-catch.cpp`  | Inner and outer try/catch with different types     |
 | `rethrow.cpp`           | Catch → modify state → rethrow (`resume` lowering) |
 | `cross-function.cpp`    | Exception propagating through multiple call levels |
-| `error-return-value.cpp`| Golden reference: hand-written lowered form        |
-
-## SAW verification
-
-`verify-lowered.saw` verifies the golden reference and (once the pass is
-implemented) the actual lowered output:
-
-- Success path: function returns correct value when no exception.
-- Error path: function returns the caught value (or sentinel) when an
-  exception is thrown.
 
 ## Running
 
 ```bash
-# Compile and lower all test cases:
+# Compile and lower all fixtures:
 ./run-tests.sh
-
-# Also run SAW verification on the golden reference:
-./run-tests.sh --verify
 
 # Clean generated artifacts:
 ./run-tests.sh --clean
@@ -39,9 +33,8 @@ implemented) the actual lowered output:
 ### Prerequisites
 
 - `clang++` with `-emit-llvm` support
-- `exception-lower` tool built from the parent directory
+- `exception-lower` tool built in `../build/`
 - `llvm-dis` (optional, for inspecting lowered IR)
-- `saw` (only needed with `--verify`)
 
 ## How the lowered form works
 
@@ -56,5 +49,6 @@ ErrorState { flag: bool, type_id: int, value: int64 }
 - **catch** → read type/value + clear flag
 - **resume** → re-set flag + return sentinel
 
-See `error-return-value.cpp` for a complete hand-written example of this
-pattern applied to the `simple-throw.cpp` test case.
+A hand-written reference of this pattern (used by the integration test)
+lives in
+[`intTests/test_exception_lower/error-return-value.cpp`](../../../intTests/test_exception_lower/error-return-value.cpp).
